12/27 장진성

+++ src/main/java/com/green/ca2sa/cafe/model/CafeGetDistanceReq.java
@@ -5,6 +5,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.beans.ConstructorProperties;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -12,11 +13,11 @@ public class CafeGetDistanceReq {
     @Schema(name="cafe_id")
     private long cafeId;
     @Schema(name="user_latitude")
-    private double userLatitude;
+    private BigDecimal userLatitude;
     @Schema(name="user_longitude")
-    private double userLongitude;
+    private BigDecimal userLongitude;
     @ConstructorProperties({"cafe_id", "user_latitude", "user_longitude"})
-    public CafeGetDistanceReq(long cafeId, double userLatitude, double longitude) {
+    public CafeGetDistanceReq(long cafeId, BigDecimal userLatitude, BigDecimal longitude) {
         this.cafeId = cafeId;
         this.userLatitude = userLatitude;
         this.userLongitude = longitude;
