+++ src/main/java/com/green/ca2sa/cafe/model/CafePutReq.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
@@ -17,6 +19,6 @@ public class CafePutReq {
     private String closeTime;
     private String openTime;
     private long cafeId;
-    private double latitude;
-    private double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
 }
